code_and_documentation/requirements.txt
PyYAML>=6.0
numpy>=1.26
matplotlib>=3.8
torch>=2.5
torchaudio>=2.5
pesq>=0.0.4
pystoi>=0.4.1
mlflow>=2.20
huggingface_hub>=0.28
speechbrain>=1.0.0
